Convert project to Maven :)
