+++ README.md
@@ -1,0 +1,5 @@
+AnimatedMOTD
+============
+Maven Dependency Version
+
+AnimatedMOTD is a BungeeCord plugin allowing you to animate the MOTD.
