+++ src/main/java/me/bigteddy98/animatedmotd/bungee/Main.java
