+++ src/main/java/me/bigteddy98/animatedmotd/bungee/Scroller.java
