+++ src/main/java/me/bigteddy98/animatedmotd/bungee/examples/AnimatedFavicon.java
@@ -38,7 +38,6 @@ public class AnimatedFavicon implements StatusListener {
 	}
 	private int currentImage = 0;
 
-	@Override
 	public ServerData update() {
 		int nextImage = this.currentImage++;
 		if (nextImage >= images.size()) {
