+++ src/main/java/me/bigteddy98/animatedmotd/bungee/ping/StatusListener.java
